Make 2 functions to avoid unnecessary newlines
Make new function we can call when we are actioned as "char"
that is NOT linewise. In this function we avoid temp files
and use printf to pipe the text to clip.exe *without* unwanted
newline at the end.

Refactor old method into its own function, and call that when
we are actioned as "line", or from visual mode.